requirements.txt
discord.py
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread==3.7.0
pytz
oauth2client
numpy
matplotlib
pynacl>=1.5.0
